Move useEventsFilter functionality to EventTabs and refactor EventTabs
