+++ src/components/EventTabs.spec.tsx
@@ -6,14 +6,11 @@ import {
 } from '../types/zetkin';
 
 describe('EventTabs', () => {
-    let dummyTimeRange : {
-        later: ZetkinEvent[] | undefined;
-        today: ZetkinEvent[] | undefined;
-        tomorrow: ZetkinEvent[] | undefined;
-        week: ZetkinEvent[] | undefined;
-    };
     let dummyEventResponses : ZetkinEventResponse[];
+    let dummyEvents : ZetkinEvent[];
     let dummyBookedEvents : ZetkinEvent[];
+
+    const today = new Date();
 
     before(()=> {
         cy.fixture('dummyEventResponses.json')
@@ -24,28 +21,36 @@ describe('EventTabs', () => {
             .then((data : {data: ZetkinEvent[]}) => {
                 dummyBookedEvents = data.data;
             });
-    });
-
-    beforeEach(()=> {
         cy.fixture('dummyEvents.json')
             .then((data : {data: ZetkinEvent[]}) => {
-                dummyTimeRange = {
-                    later: data.data,
-                    today: data.data,
-                    tomorrow: data.data,
-                    week: data.data,
-                };
+                dummyEvents = data.data;
             });
     });
 
-    it('contains different tabs for time filtering', () => {
+    it('contains tabs for time filtering on today and whis week', () => {
+        const filterDummyEvents = [
+            { ...dummyEvents[0] },
+            { ...dummyEvents[0] },
+            { ...dummyEvents[0] },
+        ];
+
+        const tomorrow = new Date();
+        tomorrow.setDate(tomorrow.getDate()+1);
+
+        const later = new Date();
+        later.setDate(later.getDate()+8);
+
+        filterDummyEvents[0].start_time = today.toISOString();
+        filterDummyEvents[1].start_time = tomorrow.toISOString();
+        filterDummyEvents[2].start_time = later.toISOString();
+
         mountWithProviders(
             <EventTabs
                 bookedEvents={ dummyBookedEvents }
                 eventResponses={ dummyEventResponses }
+                events={ filterDummyEvents }
                 onSignup={ () => null  }
                 onUndoSignup={ () => null  }
-                timeRange={ dummyTimeRange }
             />,
         );
 
@@ -55,14 +60,17 @@ describe('EventTabs', () => {
         cy.contains('misc.eventTabs.tabs.later');
     });
 
+
     it('contains event previews', () => {
+        dummyEvents[0].start_time = today.toISOString();
+
         mountWithProviders(
             <EventTabs
                 bookedEvents={ dummyBookedEvents }
                 eventResponses={ dummyEventResponses }
+                events={ dummyEvents }
                 onSignup={ () => null  }
                 onUndoSignup={ () => null  }
-                timeRange={ dummyTimeRange }
             />,
         );
