+++ src/hooks/index.ts
@@ -115,49 +115,3 @@ export const useUserFollowing = () : UserFollowing => {
 
     return { following, onFollow, onUnfollow };
 };
-
-type EventsFilter = {
-    later: ZetkinEvent[] | undefined;
-    today: ZetkinEvent[] | undefined;
-    tomorrow: ZetkinEvent[] | undefined;
-    week: ZetkinEvent[] | undefined;
-}
-
-export const useEventsFilter = (events : ZetkinEvent[] | undefined) : EventsFilter => {
-    function extractISODate(string : string) {
-        return string.slice(0, 10);
-    }
-
-    function createDateString(increment? : number) {
-        const today = new Date();
-
-        if (increment) {
-            today.setDate(today.getDate() + increment);
-        }
-
-        const date = extractISODate(today.toISOString());
-        return date;
-    }
-
-    const today = events?.filter(event =>
-        extractISODate(event.start_time) === createDateString());
-
-    const tomorrow = events?.filter(event =>
-        extractISODate(event.start_time) === createDateString(1));
-
-    const week = events?.filter(event =>
-        extractISODate(event.start_time) <= createDateString(7)
-        && extractISODate(event.start_time) >= createDateString());
-
-    const later = events?.filter(event =>
-        extractISODate(event.start_time) > createDateString(7));
-
-    const timeRange = {
-        later,
-        today,
-        tomorrow,
-        week,
-    };
-
-    return timeRange;
-};
